+++ src/test/java/pl/smsapi/api/action/subusers/SubusersJsonMother.java
@@ -1,0 +1,14 @@
+package pl.smsapi.api.action.subusers;
+
+public class SubusersJsonMother {
+
+    public static String create() {
+        return
+            "{" +
+            "  \"size\": 1," +
+            "  \"collection\": [" +
+            SubuserJsonMother.create() +
+            "  ]" +
+            "}";
+    }
+}