+++ src/test/java/pl/smsapi/test/doubles/ActionMock.java
@@ -1,15 +1,16 @@
 package pl.smsapi.test.doubles;
 
 import pl.smsapi.api.action.AbstractAction;
+import pl.smsapi.api.response.Response;
 
-public class ActionMock extends AbstractAction {
+public class ActionMock extends AbstractAction<Response> {
     @Override
     protected String endPoint() {
         return null;
     }
 
     @Override
-    protected Object createResponse(String data) {
+    protected Response createResponse(String data) {
         return null;
     }